requirements.txt
openai>=1.44.0
python-dotenv>=1.0.1
tqdm>=4.66.5
pypdf>=5.0.0
